+++ layers/Schema/packages/categories/src/ModuleProcessors/CategoryFilterInputContainerModuleProcessor.php
@@ -14,19 +14,24 @@ class CategoryFilterInputContainerModuleProcessor extends AbstractFilterInputCon
 
     public const MODULE_FILTERINPUTCONTAINER_CATEGORIES = 'filterinputcontainer-categories';
     public const MODULE_FILTERINPUTCONTAINER_CATEGORYCOUNT = 'filterinputcontainer-categorycount';
+    public const MODULE_FILTERINPUTCONTAINER_CHILDCATEGORIES = 'filterinputcontainer-childcategories';
+    public const MODULE_FILTERINPUTCONTAINER_CHILDCATEGORYCOUNT = 'filterinputcontainer-childcategorycount';
 
     public function getModulesToProcess(): array
     {
         return array(
             [self::class, self::MODULE_FILTERINPUTCONTAINER_CATEGORIES],
             [self::class, self::MODULE_FILTERINPUTCONTAINER_CATEGORYCOUNT],
+            [self::class, self::MODULE_FILTERINPUTCONTAINER_CHILDCATEGORIES],
+            [self::class, self::MODULE_FILTERINPUTCONTAINER_CHILDCATEGORYCOUNT],
         );
     }
 
     public function getFilterInputModules(array $module): array
     {
-        return match ($module[1]) {
-            self::MODULE_FILTERINPUTCONTAINER_CATEGORIES => [
+        $filterInputModules = match ($module[1]) {
+            self::MODULE_FILTERINPUTCONTAINER_CATEGORIES,
+            self::MODULE_FILTERINPUTCONTAINER_CHILDCATEGORIES => [
                 [CommonFilterInputModuleProcessor::class, CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_SEARCH],
                 [CommonFilterInputModuleProcessor::class, CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_ORDER],
                 [CommonFilterInputModuleProcessor::class, CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_LIMIT],
@@ -35,7 +40,8 @@ class CategoryFilterInputContainerModuleProcessor extends AbstractFilterInputCon
                 [CommonFilterInputModuleProcessor::class, CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_ID],
                 [FilterInputModuleProcessor::class, FilterInputModuleProcessor::MODULE_FILTERINPUT_SLUGS],
             ],
-            self::MODULE_FILTERINPUTCONTAINER_CATEGORYCOUNT => [
+            self::MODULE_FILTERINPUTCONTAINER_CATEGORYCOUNT,
+            self::MODULE_FILTERINPUTCONTAINER_CHILDCATEGORYCOUNT => [
                 [CommonFilterInputModuleProcessor::class, CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_SEARCH],
                 [CommonFilterInputModuleProcessor::class, CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_IDS],
                 [CommonFilterInputModuleProcessor::class, CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_ID],
@@ -43,6 +49,13 @@ class CategoryFilterInputContainerModuleProcessor extends AbstractFilterInputCon
             ],
             default => [],
         };
+        if (in_array($module[1], [
+            self::MODULE_FILTERINPUTCONTAINER_CATEGORIES,
+            self::MODULE_FILTERINPUTCONTAINER_CATEGORYCOUNT,
+        ])) {
+            $filterInputModules[] = [CommonFilterInputModuleProcessor::class, CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_PARENT_ID];
+        }
+        return $filterInputModules;
     }
 
     /**