+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/CHANGELOG.md
@@ -46,6 +46,10 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Filter users by email:
   - `Root.unrestrictedUsers: [User]!` ("admin" field)
   - `Root.unrestrictedUserCount: Int!` ("admin" field)
+- Added fields for Categories:
+  - `PostCategory.childCategories: [PostCategory]!`
+  - `PostCategory.childCategoryNames: [String]!`
+  - `PostCategory.childCategoryCount: Int`
 - Added fields for Menus:
   - `Root.menus: [Menu]!`
   - `Root.menuByLocation: Menu`